Registry project (folder containing Registry tables).

Account$list_registry_projects() is the typical way to get a RegistryProject.

See also

Public fields

id

String. Unique identifier.

Methods

Public methods

RegistryProject$new()

RegistryProject$print()

RegistryProject$load()

RegistryProject$get_display_name()

RegistryProject$list_tables()

RegistryProject$clone()

Method new()

Get the representation of a specific project.

Usage

RegistryProject$new(id)

Arguments

id

String. The ID of the target project.

Method print()

Print a human-readable representation of this project.

Usage

RegistryProject$print()

Method load()

(Re-)populate this project instance's cache.

Future calls to most getters will return immediately without making a network request.

Always makes a network request.

Usage

RegistryProject$load()

Method get_display_name()

Get the display name of this project.

This is an opaque string that can contain any valid Unicode data.

Display names are not unique and must never be used as identifiers. Use id instead.

Usage

RegistryProject$get_display_name(load_if_missing = TRUE)

Arguments

load_if_missing

Logical.
If TRUE, load() the display name if not in cache. If FALSE, return NULL if not in cache.

Returns

String value of the display name.

Method list_tables()

List tables contained in this project.

Usage

RegistryProject$list_tables(load_if_missing = TRUE)

Arguments

load_if_missing

Logical.
If TRUE, load() the table list if not in cache. If FALSE, return NULL if not in cache.

Returns

List of tables contained in this project.

Method clone()

The objects of this class are cloneable with this method.

Usage

RegistryProject$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.