Registry table. Contains records.

RegistryProject$list_tables() is the typical way to get a RegistryTable.

See also

Public fields

id

String. Unique identifier.

Methods

Public methods

RegistryTable$new()

RegistryTable$print()

RegistryTable$load()

RegistryTable$get_display_name()

RegistryTable$get_columns()

RegistryTable$list_records()

RegistryTable$clone()

Method new()

Get the representation of a specific table.

Usage

RegistryTable$new(id)

Arguments

id

String. The ID of the target table.

Method print()

Print a human-readable representation of this table.

Usage

RegistryTable$print()

Method load()

(Re-)populate this table instance's cache.

Future calls to most getters will return immediately without making a network request.

Always makes a network request.

Usage

RegistryTable$load()

Method get_display_name()

Get the display name of this table.

This is an opaque string that can contain any valid Unicode data.

Display names are not unique and must never be used as identifiers. Use id instead.

Usage

RegistryTable$get_display_name(load_if_missing = TRUE)

Arguments

load_if_missing

Logical.
If TRUE, load() the display name if not in cache. If FALSE, return NULL if not in cache.

Returns

String value of the display name.

Method get_columns()

Get the columns of this table.

Usage

RegistryTable$get_columns(load_if_missing = TRUE)

Arguments

load_if_missing

Logical.
If TRUE, load() the columns if not in cache. If FALSE, return NULL if not in cache.

Returns

Named list of columns of this table.

Method list_records()

List records contained in this table.

Usage

RegistryTable$list_records(page_size = 100)

Arguments

page_size

Number of records to return at a time

load_if_missing

Logical.
If TRUE, load() the records if not in cache. If FALSE, return NULL if not in cache.

Returns

coro::generator() returning pages of records from this table.

Method clone()

The objects of this class are cloneable with this method.

Usage

RegistryTable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.